pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "pydantic-ai>=1.0.9",
    "pydantic-graph>=1.0.9",
    "python-dotenv>=1.0.0",
]
